move python parsing code under dedicated package
